requirements.txt
# JIRA-Lark Base 同步系統依賴包
# Python 版本要求: >= 3.8

# Web 框架和 API
Flask>=2.3.0,<3.0.0
Flask-CORS>=4.0.0,<5.0.0

# HTTP 客戶端和認證
requests>=2.28.0,<3.0.0

# 配置文件處理
PyYAML>=6.0,<7.0
ruamel.yaml>=0.18.0,<1.0.0  # 保留註解和格式的 YAML 處理

# 文件系統監控 (配置熱重載)
watchdog>=3.0.0,<4.0.0

# 開發和除錯依賴 (可選)
# 取消註解以下行來安裝開發工具
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0